AISA: overhaul pocl image process thread
This commit does a number of things,

* separates imagereader and previewsurface
as they were sharing the same image buffer
and pocl kept a too long hold on it,
preventing the previewsurface to update.

* no more runnables on the pocl thread.
Instead, one function is called on a
separate thread that manages setup, execution
and destruction of the pocl objects.

* this thread also requests images to be
captured in switching buffer style. Like
this, the camera can make a capture while
pocl is working on the previous image.