AISA: connect imagereader to repeating request
by doing this, the imagereader is again connected
the firehose that is the repeating request. So
there are now 30 images per second generated for the
imagereader, so we need a big enough buffer to match
match the rate at which we consume these.

the benefit of this is that the images are very recent.
by using a semaphore, we can sync the image process loop
with the images being generated .